Split of image processing functions to separate file improcess.py
