+++ neossat/__init__.py
@@ -1,1 +1,2 @@
 from .neossatlib import *  # TODO might make this more complicated it we decide to split neossatlib a bit.
+from .improcess import *
